Added search parameters and hard-coded dummy search results
Created the search parameters view in order to change parameters in a search and
hard-coded some search results in some dummy view also, that is not supposed to
go to production since we agreed on using cards for the search results. Also
created some resource files that were needed for this. Since we work very
sparingly, I thought I should checkin this code even thought it's not production
ready, if any problems with this talk to me. Hope at least you guys are seeing
the commits.