Added search results view
Summary: Added hard-coded models. Changed the management of fragments by the main activity, now the fragments that are responsible for their tab labels. Also had to create a loader for the search fragment to return with network results. Created the view with a GridView along with its adapter. Created the view for the doctor's card (ProfessionalCardView) and the RatingView, which handles the rating stars. Also changed the search parameters design, now they are at the bottom and are green. Extracted some styles and strings.

Test Plan: Open up application, make sure the search fragment is displayed, make sure images are loaded with internet connection, scroll through the list and make sure the other images are loaded. Make sure cards are properly displayed.

Reviewers: samuel, estima

Reviewed By: estima

Differential Revision: http://ec2-54-186-131-188.us-west-2.compute.amazonaws.com/D18