Update/add models and add libraries for future use with search parameters
Summary: Added libraries that will be needed for supporting the search parameter views. Modified some models and added Insurance and Specialty. Also modified the Hardcoded class to include them and created a script within it to generate Ruby files for the hardcoded server db. The rest are minor changes.

Test Plan: Does not change the application behavior.

Reviewers: samuel, estima

Reviewed By: estima

Differential Revision: http://ec2-54-186-131-188.us-west-2.compute.amazonaws.com/D26