Parameters fragments integration into SearchFragment
Summary: Integrate the call to the fragments from within SearchFragment. Also had to create a factory for creating fragments on-the-fly and not sticking to them, in onder to allow the framework to claim their memory and GC to collect them. Also created some logic for the loading state in SearchFragment.

Test Plan: Verify we load search results on the search page. Change all parameters and verify results change.

Reviewers: samuel, estima

Reviewed By: estima

Differential Revision: http://ec2-54-186-131-188.us-west-2.compute.amazonaws.com/D30