Make ListenableFuture cancelable
Summary: Created a cancelable callable in order to allow tasks to cancel themselves. Also created a decorator/wrapper for the listening executor service and executor service from jdk. This is supposed to be used in hanging connections that are not interruptible from thread interrupt signals.

Test Plan: Does not change the behavior of the application.

Reviewers: samuel, estima

Reviewed By: estima

Differential Revision: http://ec2-54-186-131-188.us-west-2.compute.amazonaws.com/D31