Fix scroll stress bug
Summary: This is not a fix itself, it's rather a workaround. The bug is that if we scroll too much on the search view while the images are not loaded the app crashes. The reason is that getInputStream() from HttpUrlConnection is not interruptible from thread interrupt signals. Thus, it hangs until it gets an answer, and if we are on a slow internet connection, subsequent requests will pile up on the queue, eventually exceeding its limit (no matter which is). This workaround uses the fact that when we scroll we trigger same requests as before, thus it reuses the old ones, storing them in map, this way we won't reach the queue threashold. But this is not a fix, since if we had too much different requests we would get the same bug.

Test Plan: Scroll through the search page back and forth indefinetely and make sure app does not crash.

Reviewers: samuel, estima

Reviewed By: estima

Differential Revision: http://ec2-54-186-131-188.us-west-2.compute.amazonaws.com/D32